move js/* to app/components for organization
