+++ app/App.jsx
@@ -5,8 +5,8 @@ import { Route, Switch } from 'react-router-dom';
 import type { Match } from 'react-router-dom';
 // import { Provider } from 'react-redux'; // redux setup
 // import store from './store'; // redux setup
-import AutomaticBlog from './AutomaticBlog';
-import WrittenBlog from './WrittenBlog';
+import AutomaticBlog from './components/AutomaticBlog';
+import WrittenBlog from './components/WrittenBlog';
 // import AsyncRoute from './AsyncRoute';
 // import preload from '../data.json';
