+++ app/components/__tests__/reducers.spec.js
