+++ app/components/automatic_blog/Graph.jsx
