+++ app/components/automatic_blog/Legend.jsx
