+++ app/components/shared/PanningLink.jsx
