+++ app/components/written_blog/PostCard.jsx
